Added sudo.sh,dec2bin,gcd,power,primeMillerRabin
sudo.sh checks if the script was run as root.
decimal2binary.cpp converts decimal number to binary string.
gcd.cpp calculates gcd of two numbers.
power.cpp calculates (a^e)%mod
primeMillerRabin.cpp checks if the number is prime or not (with a
probability depending on number of iterations).